1930. Unique Length-3 Palindromic Subsequences
1930. Unique Length-3 Palindromic Subsequences
- Runtime 134 ms - ?%
- Memory 40.24 MB - ?%

+++ src/main/kotlin/t_t_se_e_z_Number_Ways_Split_Array/NumberWaysSplitArray.kt
@@ -29,7 +29,7 @@ class NumberWaysSplitArray {
             if (leftSum[i] >= rightSum[i + 1]) sum++
         }
 
-        return sum
+        return sum.toInt()
     }
 }
